+++ Taller/Frac.c
@@ -2,7 +2,7 @@
 #include <math.h>
 #include "Frac.h"
 
-long double Frac( long double x)
+ double Frac(  double x)
 {
     return x-floor(x);
 }
