+++ Taller/GHAMatrix.h
@@ -1,6 +1,6 @@
 #ifndef GHAMATRIX_H_INCLUDED
 #define GHAMATRIX_H_INCLUDED
 
-void ghaMatrix (long double Mjd_UT1, long double matrizTransformada[3][3], long double (*eop)[13]);
+void ghaMatrix ( double Mjd_UT1,  double matrizTransformada[3][3],  double (*eop)[13]);
 
 #endif // GHAMATRIX_H_INCLUDED
