+++ Taller/Gibbs.h
@@ -1,11 +1,11 @@
 #ifndef GIBBS_H_INCLUDED
 #define GIBBS_H_INCLUDED
 
-long double gibbs (long double r1[3],
-              long double r2[3],
-              long double r3[3],
-              long double vectVelocidadSal[3],
-              long double angulos[2],
+ double gibbs ( double r1[3],
+               double r2[3],
+               double r3[3],
+               double vectVelocidadSal[3],
+               double angulos[2],
               char *error
              );
