+++ Taller/R_z.h
@@ -1,6 +1,6 @@
 #ifndef R_Z_H_INCLUDED
 #define R_Z_H_INCLUDED
 
-void R_z(long double angle, long double matrizRotada[3][3]);
+void R_z( double angle,  double matrizRotada[3][3]);
 
 #endif // R_Z_H_INCLUDED
