+++ Taller/angl.h
@@ -1,6 +1,6 @@
 #ifndef ANGL_H_INCLUDED
 #define ANGL_H_INCLUDED
 
-long double angl(long double vec1[], long double vec2[]);
+ double angl( double vec1[],  double vec2[]);
 
 #endif // ANGL_H_INCLUDED
